Properly handle getting/updating fields for different object types. Squashed commit of the following:
commit 7e9ee84159c9b45781f86423c9ddb66d6f8eaf05
Author: Justin Sternberg <justin@dsgnwrks.pro>
Date:   Thu Oct 20 17:35:08 2016 -0400

    Fix bleeding global scope by resetting object params at teardown, and uncomment tests

commit 41c542277b6c9a2c730793f0b8507e178d522fb1
Author: Justin Sternberg <justin@dsgnwrks.pro>
Date:   Thu Oct 20 17:31:09 2016 -0400

    Make CMB2_REST::$boxes and CMB2_REST::$type_boxes properties protected

    Also add a get_all method for $boxes

commit 4e6044d6bd9a3a400286a4c7cf58b10dc93f39df
Merge: 82a6679 4a91731
Author: Justin Sternberg <justin@dsgnwrks.pro>
Date:   Wed Oct 19 11:46:16 2016 -0400

    Merge branch 'trunk' into cmb2-rest-api-update

commit 82a6679ddcc1b3084c49ee125ed23333793478a6
Author: Justin Sternberg <justin@dsgnwrks.pro>
Date:   Wed Oct 19 09:53:36 2016 -0400

    Need to figure out what the deal is with the tests initiations since they run in isolation, but not with all tests

+++ tests/cmb-rest-tests-base.php
@@ -43,19 +43,32 @@ abstract class Test_CMB2_Rest_Base extends Test_CMB2 {
 		foreach ( $this->metabox_array['fields'] as $field ) {
 			update_post_meta( $this->post_id, $field['id'], md5( $field['id'] ) );
 		}
+
+		CMB2_REST::register_cmb2_fields();
 	}
 
 	public function tearDown() {
 		parent::tearDown();
-		foreach ( $this->metabox_array['fields'] as $field ) {
-			delete_post_meta( $this->post_id, $field['id'] );
+		if ( ! empty( $this->metabox_array['fields'] ) ) {
+			foreach ( $this->metabox_array['fields'] as $field ) {
+				delete_post_meta( $this->post_id, $field['id'] );
+			}
 		}
-		CMB2_Boxes::remove( $this->cmb_id );
-		CMB2_REST::remove( $this->cmb_id );
+
+		Test_CMB2_REST_Object::reset_boxes();
+		Test_CMB2_REST_Object::reset_type_boxes();
+
+		foreach ( CMB2_Boxes::get_all() as $box ) {
+			CMB2_Boxes::remove( $box->cmb_id );
+		}
+
+		global $wp_actions, $wp_rest_server;
+		unset( $wp_rest_server );
+		unset( $wp_actions['rest_api_init'] );
 	}
 
 	protected function reset_instances() {
-		foreach ( CMB2_REST::$boxes as $cmb_id => $rest ) {
+		foreach ( CMB2_REST::get_all() as $cmb_id => $rest ) {
 			$rest = new CMB2_REST( $rest->cmb );
 			$rest->universal_hooks();
 		}
@@ -116,7 +129,7 @@ abstract class Test_CMB2_Rest_Base extends Test_CMB2 {
 
 if ( ! class_exists( 'Test_CMB2_REST_Object' ) ) {
 	/**
-	 * Simply allows access to the mb_defaults protected property (for testing)
+	 * Make some things accessible in the CMB2_REST class.
 	 */
 	class Test_CMB2_REST_Object extends CMB2_REST {
 		public function declare_read_edit_fields() {
@@ -128,8 +141,19 @@ if ( ! class_exists( 'Test_CMB2_REST_Object' ) ) {
 		public function can_edit( $show_in_rest ) {
 			return parent::can_edit( $show_in_rest );
 		}
-		public static function get_object_data( $object ) {
-			return parent::get_object_data( $object );
+		public static function get_object_id( $object, $object_type = 'post' ) {
+			return parent::get_object_id( $object, $object_type );
+		}
+		public static function reset_boxes() {
+			parent::$boxes = array();
+		}
+		public static function reset_type_boxes() {
+			parent::$type_boxes = array(
+				'post' => array(),
+				'user' => array(),
+				'comment' => array(),
+				'term' => array(),
+			);
 		}
 	}
 }
